# packages/api-client/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Aperture API",
        "version": "0.1.0",
        "description": "Aperture's HTTP API. This schema is the single source of truth for the frontend's types: it generates packages/api-client, which is never hand-edited."
    },
    "paths": {
...
        "/api/posts/": {
            "post": {
                "operationId": "posts_create",
                "description": "Publish a post from media that has finished processing.",
                "tags": [
                    "posts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreatePostRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/{post_id}": {
            "get": {
                "operationId": "posts_retrieve",
                "description": "One post.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            },
            "delete": {
                "operationId": "posts_destroy",
                "description": "Soft delete your own post.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/{post_id}/comments": {
            "get": {
                "operationId": "posts_comments_list",
                "description": "Top-level comments, oldest first.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Snowflake id of the last item on the previous page. Ids are time-ordered, so this is simply 'older than that one'."
                    },
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentPage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            },
            "post": {
                "operationId": "posts_comments_create",
                "description": "Comment on a post, or reply to a comment.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommentRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommentRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCommentRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Comment"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/{post_id}/like": {
            "post": {
                "operationId": "posts_like",
                "description": "Like a post. Idempotent.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            },
            "delete": {
                "operationId": "posts_unlike",
                "description": "Remove a like. Idempotent.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "post_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/by/{username}": {
            "get": {
                "operationId": "posts_by_user",
                "description": "One account's posts, newest first.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Snowflake id of the last item on the previous page. Ids are time-ordered, so this is simply 'older than that one'."
                    },
                    {
                        "in": "path",
                        "name": "username",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPage"
                                }
                            }
                        },
                        "description": ""
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/comments/{comment_id}": {
            "delete": {
                "operationId": "posts_comments_destroy",
                "description": "Soft delete your own comment.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "comment_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]+$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    },
                    "404": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/posts/feed": {
            "get": {
                "operationId": "posts_feed",
                "description": "Posts from accounts you follow, newest first.",
                "parameters": [
                    {
                        "in": "query",
                        "name": "cursor",
                        "schema": {
                            "type": "string"
                        },
                        "description": "Snowflake id of the last item on the previous page. Ids are time-ordered, so this is simply 'older than that one'."
                    }
                ],
                "tags": [
                    "posts"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostPage"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Comment": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "post_id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "parent_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "readOnly": true
                    },
                    "author": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "readOnly": true
                    },
                    "body": {
                        "type": "string",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "reply_count": {
                        "type": "integer",
                        "readOnly": true
                    }
                },
                "required": [
                    "author",
                    "body",
                    "created_at",
                    "id",
                    "parent_id",
                    "post_id",
                    "reply_count"
                ]
            },
            "CommentPage": {
                "type": "object",
                "properties": {
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Comment"
                        },
                        "readOnly": true
                    },
                    "next_cursor": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "comments",
                    "next_cursor"
                ]
            },
            "CreateCommentRequest": {
                "type": "object",
                "properties": {
                    "body": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 2200
                    },
                    "parent_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "minLength": 1
                    }
                },
                "required": [
                    "body"
                ]
            },
            "CreatePostRequest": {
                "type": "object",
                "description": "Publishing a post from media that is already uploaded and processed.",
                "properties": {
                    "media_ids": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1
                        },
                        "maxItems": 10,
                        "minItems": 1
                    },
                    "caption": {
                        "type": "string",
                        "default": "",
                        "maxLength": 2200
                    },
                    "location": {
                        "type": "string",
                        "default": "",
                        "maxLength": 120
                    },
                    "visibility": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/VisibilityEnum"
                            }
                        ],
                        "default": "public"
                    }
                },
                "required": [
                    "media_ids"
                ]
            },
...
            "Post": {
                "type": "object",
                "description": "A post as the feed and the contact sheet see it.",
                "properties": {
                    "id": {
                        "type": "string",
                        "readOnly": true
                    },
                    "author": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            }
                        ],
                        "readOnly": true
                    },
                    "caption": {
                        "type": "string",
                        "readOnly": true
                    },
                    "location": {
                        "type": "string",
                        "readOnly": true
                    },
                    "visibility": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/VisibilityEnum"
                            }
                        ],
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "media": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Media"
                        },
                        "readOnly": true
                    },
                    "like_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "comment_count": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "viewer_has_liked": {
                        "type": "boolean",
                        "readOnly": true
                    }
                },
                "required": [
                    "author",
                    "caption",
                    "comment_count",
                    "created_at",
                    "id",
                    "like_count",
                    "location",
                    "media",
                    "viewer_has_liked",
                    "visibility"
                ]
            },
            "PostPage": {
                "type": "object",
                "description": "A cursor-paginated page of posts.\n\n`next_cursor` is the id of the last post on the page, or null at the end.\nNo total count: producing one means a `COUNT(*)` over the whole feed, and\nnothing in the UI shows it.",
                "properties": {
                    "posts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Post"
                        },
                        "readOnly": true
                    },
                    "next_cursor": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "next_cursor",
                    "posts"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "aperture_session"
            }
        }
    }
}